Adds Travis config for continuous integration
The config builds and tests the application as normal, and if it
succeeds it will attempt to deploy to the Github Pages staging "server".

The config securely stores a GH access token, and uses this to force
push the (usually) ignored asset files, like `js/bundle.js`. GH Pages
will then build the Jekyll assets, along with the Javascripts, and start
serving the latest changes.